validator/requirements.txt
fastapi
keycloak-client
uvicorn
requests
